Strength typings, second pass.
