+++ docs/lavalink.rst
@@ -37,9 +37,6 @@ Errors
 .. autoclass:: RequestError
     :members:
 
-.. autoclass:: PlayerErrorEvent
-    :members:
-
 Events
 ------
 All Events are derived from :class:`Event`
@@ -84,6 +81,9 @@ All Events are derived from :class:`Event`
     :members:
 
 .. autoclass:: IncomingWebSocketMessage
+    :members:
+
+.. autoclass:: PlayerErrorEvent
     :members:
 
 Filters
